Extract information from S&P directly from HTML
* Create helper file for fetching
* Reorganize fetchers and related interfaces
* Create method for capturing fetch errors for every fetcher type
