+++ .yarnrc.yml
@@ -1,3 +1,9 @@
+compressionLevel: mixed
+
+defaultSemverRangePrefix: ""
+
+networkConcurrency: 1024
+
 packageExtensions:
   es-iterator-helpers@*:
     dependencies:
@@ -20,7 +26,3 @@ packageExtensions:
   vite-tsconfig-paths@*:
     dependencies:
       vite: "*"
-
-compressionLevel: mixed
-defaultSemverRangePrefix: ""
-networkConcurrency: 1024
