+++ packages/backend/package.json
@@ -35,6 +35,7 @@
     "@prisma/client": "5.5.2",
     "@rating-tracker/commons": "workspace:*",
     "@simplewebauthn/server": "8.3.3",
+    "@xmldom/xmldom": "0.8.10",
     "axios": "1.6.0",
     "chalk": "5.3.0",
     "cookie-parser": "1.4.6",
@@ -52,7 +53,8 @@
     "response-time": "2.3.2",
     "selenium-webdriver": "4.14.0",
     "swagger-ui-dist": "5.9.1",
-    "swagger-ui-express": "5.0.0"
+    "swagger-ui-express": "5.0.0",
+    "xpath": "0.0.33"
   },
   "devDependencies": {
     "@types/cookie-parser": "1.4.5",
