+++ packages/backend/src/fetchers/marketScreenerFetcher.ts
@@ -2,13 +2,13 @@ import { Stock } from "@rating-tracker/commons";
 import { Request } from "express";
 import { By, WebDriver, until } from "selenium-webdriver";
 
-import type { SeleniumFetcher } from "../controllers/FetchController";
-import { FetcherWorkspace } from "../controllers/FetchController";
 import { readStock, updateStock } from "../db/tables/stockTable";
 import * as signal from "../signal/signal";
 import { SIGNAL_PREFIX_ERROR } from "../signal/signal";
 import logger from "../utils/logger";
-import { openPageAndWait, takeScreenshot } from "../utils/webdriver";
+import { openPageAndWait } from "../utils/webdriver";
+
+import { type SeleniumFetcher, type FetcherWorkspace, captureFetchError } from "./fetchHelper";
 
 const XPATH_ANALYST_COUNT =
   "//div[@class='card-content']/div/div/div[contains(text(), 'Number of Analysts')]/following-sibling::div" as const;
@@ -163,8 +163,8 @@ const marketScreenerFetcher: SeleniumFetcher = async (
   });
   if (errorMessage.includes("\n")) {
     // An error occurred if and only if the error message contains a newline character.
-    // We take a screenshot and send a message.
-    errorMessage += `\n${await takeScreenshot(driver, stock, "marketScreener")}`;
+    // We capture the resource and send a message.
+    errorMessage += `\n${await captureFetchError(stock, "marketScreener", { driver })}`;
     await signal.sendMessage(SIGNAL_PREFIX_ERROR + errorMessage, "fetchError");
     stocks.failed.push(await readStock(stock.ticker));
   } else {
