+++ packages/backend/src/fetchers/msciFetcher.ts
@@ -2,12 +2,13 @@ import { MSCIESGRating, Stock, isMSCIESGRating } from "@rating-tracker/commons";
 import { Request } from "express";
 import { By, WebDriver, until } from "selenium-webdriver";
 
-import { FetcherWorkspace, SeleniumFetcher } from "../controllers/FetchController";
 import { readStock, updateStock } from "../db/tables/stockTable";
 import * as signal from "../signal/signal";
 import { SIGNAL_PREFIX_ERROR } from "../signal/signal";
 import logger from "../utils/logger";
-import { openPageAndWait, takeScreenshot } from "../utils/webdriver";
+import { openPageAndWait } from "../utils/webdriver";
+
+import { type SeleniumFetcher, type FetcherWorkspace, captureFetchError } from "./fetchHelper";
 
 /**
  * Fetches data from MSCI using a Selenium WebDriver.
@@ -104,8 +105,8 @@ const msciFetcher: SeleniumFetcher = async (
   });
   if (errorMessage.includes("\n")) {
     // An error occurred if and only if the error message contains a newline character.
-    // We take a screenshot and send a message.
-    errorMessage += `\n${await takeScreenshot(driver, stock, "msci")}`;
+    // We capture the resource and send a message.
+    errorMessage += `\n${await captureFetchError(stock, "msci", { driver })}`;
     await signal.sendMessage(SIGNAL_PREFIX_ERROR + errorMessage, "fetchError");
     stocks.failed.push(await readStock(stock.ticker));
   } else {
