+++ packages/backend/src/fetchers/refinitivFetcher.ts
@@ -1,13 +1,16 @@
+import assert from "node:assert";
+
 import { Stock } from "@rating-tracker/commons";
 import axios from "axios";
 import { Request } from "express";
 
-import { Fetcher, FetcherWorkspace } from "../controllers/FetchController";
 import { readStock, updateStock } from "../db/tables/stockTable";
 import * as signal from "../signal/signal";
 import { SIGNAL_PREFIX_ERROR } from "../signal/signal";
 import APIError from "../utils/APIError";
 import logger from "../utils/logger";
+
+import { type JSONFetcher, type FetcherWorkspace, captureFetchError } from "./fetchHelper";
 
 /**
  * Fetches data from Refinitiv.
@@ -16,26 +19,33 @@ import logger from "../utils/logger";
  * @param {FetcherWorkspace} stocks An object with the stocks to fetch and the stocks already fetched (successful or
  * with errors)
  * @param {Stock} stock The stock to extract data for
+ * @param {string} json The fetched and parsed JSON
  * @returns {boolean} Whether the driver is still healthy
  * @throws an {@link APIError} in case of a severe error
  */
-const refinitivFetcher: Fetcher = async (
+const refinitivFetcher: JSONFetcher = async (
   req: Request,
   stocks: FetcherWorkspace<Stock>,
   stock: Stock,
+  json: Object,
 ): Promise<boolean> => {
   let refinitivESGScore: number = req.query.clear ? null : undefined;
   let refinitivEmissions: number = req.query.clear ? null : undefined;
 
   const url = `https://www.refinitiv.com/bin/esg/esgsearchresult?ricCode=${stock.ric}`;
 
-  const refinitivJSON = (await axios.get(url)).data;
+  json = (await axios.get(url)).data;
 
-  if (Object.keys(refinitivJSON).length === 0 && refinitivJSON.constructor === Object) {
+  if (Object.keys(json).length === 0 && json.constructor === Object) {
     throw new APIError(502, "No Refinitiv information available.");
   }
 
-  if (refinitivJSON.status && refinitivJSON.status.limitExceeded === true) {
+  if (
+    "status" in json &&
+    typeof json.status === "object" &&
+    "limitExceeded" in json.status &&
+    json.status.limitExceeded === true
+  ) {
     // If the limit has been exceeded, we stop fetching data and throw an error.
     throw new APIError(429, "Limit exceeded.");
   }
@@ -44,7 +54,15 @@ const refinitivFetcher: Fetcher = async (
   let errorMessage = `Error while fetching Refinitiv information for stock ${stock.ticker}:`;
 
   try {
-    refinitivESGScore = +refinitivJSON.esgScore["TR.TRESG"].score;
+    assert(
+      "esgScore" in json &&
+        typeof json.esgScore === "object" &&
+        "TR.TRESG" in json.esgScore &&
+        typeof json.esgScore["TR.TRESG"] === "object" &&
+        "score" in json.esgScore["TR.TRESG"],
+      "Refinitiv ESG Score not found in JSON response.",
+    );
+    refinitivESGScore = +json.esgScore["TR.TRESG"].score;
   } catch (e) {
     logger.warn({ prefix: "selenium" }, `Stock ${stock.ticker}: Unable to extract Refinitiv ESG Score: ${e}`);
     if (stock.refinitivESGScore !== null) {
@@ -60,7 +78,15 @@ const refinitivFetcher: Fetcher = async (
   }
 
   try {
-    refinitivEmissions = +refinitivJSON.esgScore["TR.TRESGEmissions"].score;
+    assert(
+      "esgScore" in json &&
+        typeof json.esgScore === "object" &&
+        "TR.TRESGEmissions" in json.esgScore &&
+        typeof json.esgScore["TR.TRESGEmissions"] === "object" &&
+        "score" in json.esgScore["TR.TRESGEmissions"],
+      "Refinitiv Emissions Score not found in JSON response.",
+    );
+    refinitivEmissions = +json.esgScore["TR.TRESGEmissions"].score;
   } catch (e) {
     logger.warn({ prefix: "selenium" }, `Stock ${stock.ticker}: Unable to extract Refinitiv Emissions: ${e}`);
     if (stock.refinitivEmissions !== null) {
@@ -83,12 +109,13 @@ const refinitivFetcher: Fetcher = async (
   });
   if (errorMessage.includes("\n")) {
     // An error occurred if and only if the error message contains a newline character.
-    errorMessage += `\nServer response: ${JSON.stringify(refinitivJSON)}`;
+    errorMessage += `\n${await captureFetchError(stock, "refinitiv", { json })}`;
     await signal.sendMessage(SIGNAL_PREFIX_ERROR + errorMessage, "fetchError");
     stocks.failed.push(await readStock(stock.ticker));
   } else {
     stocks.successful.push(await readStock(stock.ticker));
   }
+  json = undefined;
   return true;
 };
